components/evaluation/requirements.txt
pandas==2.1.4
llama-cpp-python